refactor(kms): use DryRunOperationException from api and clean up imports
